+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Development
 - Switch from nosetest to pytest as test runner #2114
 - The codebase is now formatted using ``black``. #2109
 - In bulk write insert, the detailed error message would raise in exception.
+- ``Doc.objects.insert`` accepts kwargs to pass ``ordered``, ``bypass_document_validation``, etc. #2169
 
 Changes in 0.18.2
 =================